+++ src/Harmony/HomepageBundle/HarmonyHomepageBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Harmony\HomepageBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class HarmonyHomepageBundle extends Bundle
+{
+}